requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
google-generativeai==0.8.5
propcache==0.3.2
pyautogui==0.9.54
python-dotenv==1.1.1
pyperclip==1.9.0
psutil==5.9.8
Pillow==10.4.0

# Для Windows (опционально, добавьте вручную при необходимости):
# pywinauto==0.6.9

# Для macOS дополнительно:
# pyobjc-core==9.2
# pyobjc==9.2

# Для проверки формата
flake8==7.3.0
black==25.1.0

# Для утилиты пиксельного зонда
pynput==1.7.7
